+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Update to hanami-router 2.0.0 stable. Use it's BodyParser middleware to parse the body.
+- Support simple form data requests. Still have to figure out file uploads.
 
 ## 0.20.0
 - You can pass a filepath to `spec:` now so you no longer have to call `OpenapiFirst.load` anymore.